Update main.js
Added a rich text editor (see comment in code).
The "getText()" function extracts the edited content from the "iframe". The function is called in "text_editor.php" and the content is handed over to hte "update_szoveg()" function in "Functions-php".
Function "frameHeight()" sets the height of the editor "ifram" to 75% percent of the screen height.