Update text_editor.php
I'm using this editor: http://nicedit.com
The "Elküld" button calls the script below to get the contetnt of the textarea, and also sends a hidden form, with the extracted text to the update_szoveg function at "Functions.php", which updates it in the database. There is a response whether the query is successful or not.